requirements.txt
Keras==2.3.1
tensorflow-gpu==1.13.1
numpy==1.19.2
opencv-python==4.2.0.32
scikit-learn==0.21.2
scipy==1.4.1
Pillow	
torch==1.3.0
torchvision==0.4.1
